+++ inc/ti/type.h
@@ -52,6 +52,11 @@ int ti_type_add_method(
         ti_closure_t * closure,
         ex_t * e);
 void ti_type_remove_method(ti_type_t * type, ti_name_t * name);
+int ti_type_set_idname(
+        ti_type_t * type,
+        const char * s,
+        size_t n,
+        ex_t * e);
 int ti_type_methods_to_pk(ti_type_t * type, msgpack_packer * pk);
 int ti_type_relations_to_pk(ti_type_t * type, msgpack_packer * pk);
 int ti_type_methods_info_to_pk(