+++ src/ti/field.c
@@ -445,6 +445,12 @@ static int field__init(ti_field_t * field, ex_t * e)
         return e->nr;
     }
 
+    if (*str == '&')
+    {
+        ++str;
+        --n;
+    }
+
     if (str[n-1] == '?')
     {
         field->spec |= TI_SPEC_NILLABLE;
@@ -971,7 +977,8 @@ int ti_field_set_name(
         return e->nr;
     }
 
-    if (ti_field_by_name(field->type, name) ||
+    if (field->type->idname == name ||
+        ti_field_by_name(field->type, name) ||
         ti_method_by_name(field->type, name))
     {
         ex_set(e, EX_VALUE_ERROR,