+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# v1.3.0
+
+* Removed deprecated functions `return(..)` and `if(..)`, pr #297.
+* Added Id (`'#'`) definition for Id mapping on typed things, issue #296.
+* Added map shortcut functions, issue #298.
+  - _list_.`map_id()`: https://docs.thingsdb.net/v1/data-types/list/map_id
+  - _list_.`map_wrap(..)`: https://docs.thingsdb.net/v1/data-types/list/map_wrap
+  - _list_.`map_type(..)`: https://docs.thingsdb.net/v1/data-types/list/map_type
+  - _set_.`map_id()`: https://docs.thingsdb.net/v1/data-types/set/map_id
+  - _set_.`map_wrap(..)`: https://docs.thingsdb.net/v1/data-types/set/map_wrap
+* Change the default *deep* value back to `1` (one), issue #300.
+* Allow `&` to prefix a definition for the same deep level when wrapping a *thing*, issue #301.
+
 # v1.2.9
 
 * Fixed rename type when used with restricted things, issue #292.